requirements.txt
fastapi
uvicorn
pydantic
elasticsearch
openai
python-dotenv
numpy
pydantic-settings